Streamline xor example
